rpmfile2.c: removed cacheproc and frenc, prints to stdout directly
The idea was to serialize records, per package, so that they could
be cached and then printed from the cache.  This was intended as
a preliminary exercise to sharpen the techniques that could later
be used to deal with elf symbols.

Anyway, the <frnec.h> library is nowhere near being ready,
and <rpmcache.h> is even less so.  This is not because I was being
lazy, but rather because it's not at all clear what could be the
best way to do certain things.  Take the cache, for example.
If a cache is backed by a hard drive, and the entries are accessed
in random order, than the user should be aware that, unless he's lucky,
he or she can only retrieve about 100 cache entries per second.
So it can take 200 seconds to process 20,000 packages.  Who needs
such a cache?

So instead, this change introduces some optimizations in filename
handling: filenames are split into dirnames+basenames, and successive
dirnames are reused.