rpmfile2.c: identify hardlinks
The first hardlink in a set (which also goes first in the sorted output)
will be identified with its magic type.  The remaining hardlinks will
be identified as "hard link to first" or "hard link to /path/first",
depending on whether the first hardlink and the succeeding hardlink
are under the same dir.

Notice that, in a payload, it is the last hardlink that comes with
the data, not the first one.  So a riddle of sorts, "How to do it in
the reverse", perplexed me for quite a while.  The solution, simple
though it may seem, certifies my great ingenuity.

$ rpmfile2 git-core-2.10.4-alt1.x86_64.rpm
/usr/bin/git	100755	ELF 64-bit LSB executable, x86-64 [...]
/usr/bin/git-upload-archive	100755	hard link to git
/usr/libexec/git-core/git	100755	hard link to /usr/bin/git
/usr/libexec/git-core/git-add	100755	hard link to /usr/bin/git
/usr/libexec/git-core/mergetools/gvimdiff	100644	ASCII text
/usr/libexec/git-core/mergetools/gvimdiff2	100644	hard link to gvimdiff
/usr/libexec/git-core/mergetools/gvimdiff3	100644	hard link to gvimdiff

The rpmcpio library now performs some extra verification of hardlink
sets.  This also works well with hardlinked symlinks, a curiosity which
rpm-4.0 can readily produce.

%install
mkdir -p %buildroot%_bindir
cd %buildroot%_bindir
ln -s perl perlx
ln perlx perly
%files
%_bindir/perl?

$ rpmfile2 foo-1.0-alt0.x86_64.rpm
/usr/bin/perlx  120777  symbolic link to perl
/usr/bin/perly  120777  hard link to perlx