rpmfile2.c: identify other S_IFMT types
The list is now complete, the output is pretty close to what
file-4.xx prints.  There's still no "suid", "sgid" or "sticky" labels
in the output.  This is partially compensated by the mode column.
In file-5.xx, these labels are are delimited with comma, like in the
following example.  So identical output is elusive anyway.

$ file /tmp
/tmp: sticky, directory