rpmfile2.c: fixed filename comparison [d1,b1] cmp [d2,b2]
What a thinko!  This is slightly harder than I first thought.
If dirname lengths are the same, then the outcome of comparison
is determined by comparing dirnames or, in addition, basenames.
However, when dirname lengths are different, the comparison is
determined by the overlap which exists between the longer dirname
and the other basename.